splitting out env
